Restore Actions

+++ app/Actions/Models/Delete/AbstractDeleteAction.php
@@ -10,8 +10,9 @@ namespace App\Actions\Models\Delete;
 
 
 use App\Actions\Models\AbstractModelAction;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\Model;
-use Roelhem\Actions\Contracts\ActionContextContract;
+use Roelhem\Actions\Contracts\ActionContext;
 use Roelhem\GraphQL\Facades\GraphQL;
 
 abstract class AbstractDeleteAction extends AbstractModelAction
@@ -21,15 +22,20 @@ abstract class AbstractDeleteAction extends AbstractModelAction
      * Default create action, by calling the delete-method on the model instance.
      *
      * @param array $validArgs
-     * @param null|ActionContextContract $context
+     * @param null|ActionContext $context
      * @return Model
      * @throws
      */
-    protected function handle($validArgs = [], ?ActionContextContract $context = null)
+    protected function handle($validArgs = [], ?ActionContext $context = null)
     {
         $id = array_get($validArgs,'id');
         /** @var Model $model */
         $model = call_user_func([$this->getModelClass(), 'findOrFail'], $id);
+
+        // Check if the update ability for this model is allowed in this context.
+        if(!$context->can('delete', $model)) {
+            throw new AuthorizationException("Not allowed to delete this model.");
+        };
 
         // Delete the model
         $model->delete();
